UPSTREAM: Revert "PCI: Use preserve_config in place of pci_flags"
This reverts commit 9f82acdb380fcb30c4ebee49223eb4dd06d4d22a which is
commit 7246a4520b4bf1494d7d030166a11b5226f6d508 upstream.

This patch causes a regression in cuttlefish/crossvm boot on arm64.

The patch was part of a series that when applied will not cause a regression
but this patch was backported to the 5.10 branch by itself.

The other patches do not apply cleanly to the 5.10 branch.

Change-Id: Ic43827b09479ce94cba23f9621f1afefa743cf91
Signed-off-by: Terry Tritton <terry.tritton@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit da55e762258004b45e171689290bf20b6d9455d8)